requirements.txt
flask
psutil
Authlib
python-dotenv
requests
